+++ src/main/java/me/villagerunknown/platform/util/StringUtil.java
@@ -6,6 +6,20 @@ public class StringUtil {
 		return string.substring(0,1).toUpperCase() + string.substring(1);
 	}
 	
+	public static String capitalizeAll( String string ) {
+		String[] words = string.split(" ");
+		
+		StringBuilder sentence = new StringBuilder();
+		
+		if( words.length > 0 ) {
+			for (String word : words) {
+				sentence.append( capitalize( word ) ).append(" ");
+			} // for
+		} // if
+		
+		return sentence.toString().trim();
+	}
+	
 	public static boolean isSingular( int number ) {
 		return (number <= 1);
 	}