+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.1]
+
+### Added
+
+- Added `define` method to `Platform` class to allow mod developers to change the "villagerunknown-" prefix.
+- Added `register` method to `Platform` class to allow mod developers to insert a `PlatformMod` object directly into the registered mods list. 
+- Added `registered` method to `Platform` class to allow mod developers to check if a mod is registered.
+- Added `registration` method to `Platform` class to allow mod developers to retrieve a mod's `PlatformMod` object.
+- Added `reportConversionToLog` method to `EntityUtil` class.
+- Added `reportConversionToChat` method to `EntityUtil` class.
+- Added `registerEntity` method to `RegistryUtil` class.
+- Added `getItemGroup` method to `RegistryUtil` class.
+- Added `registerItemGroup` method to `RegistryUtil` class.
+- Added `addItemToGroup` method to `RegistryUtil` class.
+- Added `capitalizeAll` method to `RegistryUtil` class.
+- Added `LocatorUtil` class to provide methods for locating structures and biomes.
+
+### Fixed
+
+- Set pretty printing in Gson to correct formatting of files created by list.StringsList class. 
+- Resolved an issue causing an incompatibility with Custom Villager Names and Healthy Habits.
+
 ## [1.1.0]
 
 ### Added